feat: putting things together
